+++ src/common/styles/sizes.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line
+const sizes = require('less-vars-loader?camelCase!./sizes.less');
+
+export default sizes;